Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCSctpTransport-constructor-expected.txt


FAIL setRemoteDescription() with answer containing data media should initialize pc.sctp assert_equals: expected (object) null but got (undefined) undefined
FAIL setLocalDescription() with answer containing data media should initialize pc.sctp assert_equals: expected (object) null but got (undefined) undefined
